# Port to new .csproj + support netstandard 1.5