--- a/StringFormatter/Arg.tt
+++ b/StringFormatter/Arg.tt
@@ -31,10 +31,9 @@ namespace System.Text.Formatting {
         /// </summary>
         /// <param name="format">A composite format string.</param><#for(int j = 0; j < i; j++) {
 #>
-
-        /// <param name="arg<#=j#>">A value to format.</param><#}#>   
-        public void AppendFormat<#=genericParams#>(string format, T0 arg0<#for(int j = 1; j < i; j++) { #>, T<#=j#> arg<#=j#><#}#>) {
-            var args = new <#=fullName#>(__makeref(arg0)<#for(int j = 1; j < i; j++) { #>, __makeref(arg<#=j#>)<#}#>);
+        /// <param name="arg<#=j#>">A value to format.</param><#}#>
+        public unsafe void AppendFormat<#=genericParams#>(string format, T0 arg0<#for(int j = 1; j < i; j++) { #>, T<#=j#> arg<#=j#><#}#>) {
+            var args = new <#=fullName#>(Unsafe.AsPointer(ref arg0)<#for(int j = 1; j < i; j++) { #>, Unsafe.AsPointer(ref arg<#=j#>)<#}#>);
             AppendArgSet(format, ref args);
         }
 
@@ -43,8 +42,7 @@ namespace System.Text.Formatting {
         /// </summary>
         /// <param name="format">A composite format string.</param><#for(int j = 0; j < i; j++) {
 #>
-
-        /// <param name="arg<#=j#>">A value to format.</param><#}#>   
+        /// <param name="arg<#=j#>">A value to format.</param><#}#>
         public static string Format<#=genericParams#>(string format, T0 arg0<#for(int j = 1; j < i; j++) { #>, T<#=j#> arg<#=j#><#}#>) {
             var buffer = Acquire(format.Length + <#=i * 8#>);
             buffer.AppendFormat(format, arg0<#for(int j = 1; j < i; j++) { #>, arg<#=j#><#}#>);
@@ -64,15 +62,14 @@ namespace System.Text.Formatting {
 #>
 
     unsafe struct <#=fullName#> : IArgSet {
-<#for(int j = 0; j < i; j++) { #>        IntPtr t<#=j#>;
+<#for(int j = 0; j < i; j++) { #>        readonly void* t<#=j#>;
 <#}#>
-
         public int Count => <#=i#>;
 
-        public <#=simpleName#> (TypedReference t0<#for(int j = 1; j < i; j++) { #>, TypedReference t<#=j#><#}#>) {
-    <#for(int j = 0; j < i; j++) { #>        this.t<#=j#> = *(IntPtr*)&t<#=j#>;
-    <#}#>
-    }
+        public <#=simpleName#> (void* t0<#for(int j = 1; j < i; j++) { #>, void *t<#=j#><#}#>) {
+<#for(int j = 0; j < i; j++) { #>            this.t<#=j#> = t<#=j#>;
+<#}#>
+        }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public void Format (StringBuffer buffer, int index, StringView format) {