--- a/StringFormatterBench/Program.cs
+++ b/StringFormatterBench/Program.cs
@@ -4,7 +4,7 @@ using System.Text.Formatting;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Columns;
 using BenchmarkDotNet.Configs;
-using BenchmarkDotNet.Diagnostics.Windows;
+using BenchmarkDotNet.Diagnosers;
 using BenchmarkDotNet.Jobs;
 using BenchmarkDotNet.Running;