--- a/StringFormatterBench/packages.config
+++ b/StringFormatterBench/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="BenchmarkDotNet" version="0.9.6" targetFramework="net452" />
-  <package id="BenchmarkDotNet.Diagnostics.Windows" version="0.9.6" targetFramework="net452" />
-  <package id="Microsoft.Diagnostics.Tracing.TraceEvent" version="1.0.41" targetFramework="net452" />
-</packages>