--- a/Test/packages.config
+++ b/Test/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="AutoFixture" version="3.50.2" targetFramework="net45" />
-  <package id="NFluent" version="1.3.1.0" targetFramework="net45" />
-  <package id="NUnit" version="2.6.4" targetFramework="net45" />
-</packages>