… StringFormatter/Arg.cs
… StringFormatter/Properties/AssemblyInfo.cs
… StringFormatterBench/Properties/AssemblyInfo.cs
… Test/Properties/AssemblyInfo.cs