# Summary
So this is my attempt at getting this to run in CoreCLR.

Thanks to @benaadams for pointing me into the right direction w.r.t: [System.Runtime.CompilerServices.Unsafe](https://www.nuget.org/packages/System.Runtime.CompilerServices.Unsafe/)

The gist of this is:
* Remove __makeref and friends in favor of simpler, portable [System.Runtime.CompilerServices.Unsafe](https://www.nuget.org/packages/System.Runtime.CompilerServices.Unsafe/)
* Change old .csproj to new .csproj to allow targeting netstandard 1.5
* No functional changes
* All unit tests run
* New .csproj generates .nupkg on build

# What's missing

Porting the cake scripts to make this thing build properly in the CI env.

# Perf bench:

No significant change that I can see:

## Before

Method |       Jit |        Median |     StdDev | Scaled |         Min |           Max |  Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
------------- |---------- |-------------- |----------- |------- |------------ |-------------- |------- |------ |------ |------------------- |
Baseline | LegacyJit | 1,031.7042 ns | 90.6496 ns |   1.00 | 941.6965 ns | 1,274.3865 ns | 965.11 |     - |     - |             263.97 |
StringBuffer | LegacyJit |   813.9107 ns |  6.4698 ns |   0.79 | 800.7643 ns |   827.6636 ns | 185.40 |     - |     - |              52.34 |
Baseline |    RyuJit |   968.1944 ns |  8.3778 ns |   1.00 | 954.2833 ns |   988.7216 ns | 870.00 |     - |     - |             238.12 |
StringBuffer |    RyuJit |   919.3434 ns |  7.3780 ns |   0.95 | 907.2533 ns |   935.1812 ns | 218.30 |     - |     - |              61.53 |
 
Method |       Jit |      Median |    StdDev | Scaled |         Min |         Max |  Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
------------- |---------- |------------ |---------- |------- |------------ |------------ |------- |------ |------ |------------------- |
Baseline | LegacyJit | 953.5759 ns | 8.1595 ns |   1.00 | 940.4195 ns | 970.5280 ns | 627.00 |     - |     - |             178.55 |
NoAllocation | LegacyJit | 828.3974 ns | 8.8559 ns |   0.87 | 811.3848 ns | 852.2946 ns |      - |     - |     - |               0.12 |
Baseline |    RyuJit | 952.0248 ns | 8.2248 ns |   1.00 | 941.6958 ns | 970.8299 ns | 599.13 |     - |     - |             170.61 |
NoAllocation |    RyuJit | 924.8467 ns | 5.5568 ns |   0.97 | 914.4467 ns | 938.0845 ns |      - |     - |     - |               0.11 |

## After

Method |          Job |       Jit |     Mean |    Error |   StdDev |      Min |      Max | Scaled |  Gen 0 | Allocated |
------------- |------------- |---------- |---------:|---------:|---------:|---------:|---------:|-------:|-------:|----------:|
Baseline | LegacyJitX64 | LegacyJit | 908.7 ns | 6.216 ns | 5.814 ns | 896.7 ns | 917.8 ns |   1.00 | 0.1774 |     560 B |
StringBuffer | LegacyJitX64 | LegacyJit | 783.8 ns | 2.808 ns | 2.193 ns | 780.0 ns | 787.7 ns |   0.86 | 0.0429 |     136 B |
Baseline |    RyuJitX64 |    RyuJit | 941.8 ns | 3.936 ns | 3.681 ns | 938.0 ns | 951.4 ns |   1.00 | 0.1774 |     560 B |
StringBuffer |    RyuJitX64 |    RyuJit | 897.7 ns | 6.177 ns | 5.778 ns | 888.7 ns | 907.3 ns |   0.95 | 0.0429 |     136 B |
 
Method |          Job |       Jit |     Mean |    Error |   StdDev |      Min |      Max | Scaled |  Gen 0 | Allocated |
------------- |------------- |---------- |---------:|---------:|---------:|---------:|---------:|-------:|-------:|----------:|
Baseline | LegacyJitX64 | LegacyJit | 931.3 ns | 6.277 ns | 5.871 ns | 922.8 ns | 940.1 ns |   1.00 | 0.1345 |     424 B |
NoAllocation | LegacyJitX64 | LegacyJit | 810.6 ns | 5.820 ns | 5.444 ns | 801.8 ns | 818.9 ns |   0.87 |      - |       0 B |
Baseline |    RyuJitX64 |    RyuJit | 941.3 ns | 1.617 ns | 1.069 ns | 939.5 ns | 942.6 ns |   1.00 | 0.1345 |     424 B |
NoAllocation |    RyuJitX64 |    RyuJit | 899.5 ns | 2.681 ns | 2.239 ns | 894.5 ns | 903.2 ns |   0.96 |      - |       0 B |

## CoreCLR 1.0.3 benchmark:
Method |       Mean |    Error |   StdDev |        Min |        Max | Scaled |  Gen 0 | Allocated |
------------- |-----------:|---------:|---------:|-----------:|-----------:|-------:|-------:|----------:|
Baseline | 1,029.3 ns | 5.983 ns | 5.304 ns | 1,017.7 ns | 1,037.4 ns |   1.00 | 0.1106 |     352 B |
StringBuffer |   820.6 ns | 6.350 ns | 5.940 ns |   808.1 ns |   827.7 ns |   0.80 | 0.0429 |     136 B |
  
 
Method |       Mean |    Error |   StdDev |      Min |        Max | Scaled |  Gen 0 | Allocated |
------------- |-----------:|---------:|---------:|---------:|-----------:|-------:|-------:|----------:|
Baseline | 1,003.6 ns | 6.882 ns | 5.747 ns | 992.9 ns | 1,016.4 ns |   1.00 | 0.0668 |     216 B |
NoAllocation |   841.2 ns | 5.989 ns | 5.602 ns | 833.9 ns |   848.4 ns |   0.84 |      - |       0 B |