@damageboy 2017-08-17
Is there anything you wish to comment on this?
Would it make it easier if I did anything else in order to accept this PR?