@MendelMonteiro 2017-08-17
Sorry for not replying earlier but I'm on holidays all this week. One of the other maintainers will have a look and merge the PR. I'm all for it.

Ping @rverdier @rbouallou