@rverdier 2017-08-18
Hey @damageboy; sorry for the late answer. I just merged the PR and updated the package (1.0.0.8).

Thanks a lot!