@damageboy 2017-08-19
Hi,
The latest nuget doesn't seem to include a lib for netstandard 1.5 and doesn't have a dependancy on System.Runtime.CompilerServices.Unsafe...

Maybe it's not really published somehow?

Maybe a build issue of some sort?