@MendelMonteiro 2017-08-21
The build script had to be adapted to create a nuget package with the right dependencies and libs.  The 1.0.0.9 package should be good.